requirements.txt
# 基础深度学习框架
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Diffusion 模型库
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0

# COCO 数据集工具
pycocotools>=2.0.6

# 图像处理
Pillow>=9.5.0
opencv-python>=4.8.0
einops>=0.6.1

# 配置和工具
omegaconf>=2.3.0
tqdm>=4.65.0
tensorboard>=2.13.0
matplotlib>=3.7.0

# 可选：PyTorch Lightning
# pytorch-lightning>=2.0.0

